+++ packages/narrat/src/components/inventory-ui.vue
@@ -1,61 +1,37 @@
 <template>
-  <modal class="menu" @close="close" containerCssClass="inventory-modal">
-    <template v-slot:header>
-      <h3 class="title">Inventory</h3>
-    </template>
-    <template v-slot:body>
-      <div
-        class="inventory-container"
-        v-if="!chosenItem && Object.keys(itemsToDisplay).length > 0"
-      >
-        <button
-          @click="() => clickItem(item.id)"
-          class="item-display"
-          :style="getItemStyle(item.id)"
-          v-for="item in itemsToDisplay"
-          :key="item.id"
-        >
-          <h3 class="item-title">{{ getItemName(item.id) }}</h3>
-          <h3 class="item-amount">{{ item.amount }}</h3>
-        </button>
-      </div>
-      <div v-else-if="typeof chosenId === 'string'">
-        <div class="flex flex-row item-description-container">
-          <div class="flex item-left">
-            <div class="item-display" :style="getItemStyle(chosenId)"></div>
-          </div>
-          <div class="flex item-right">
-            <h2>{{ getItemName(chosenId) }}</h2>
-            <hr class="hr-solid" />
-            <h3>Amount: {{ chosenItem!.amount }}</h3>
-            <p>{{ chosenItemConf!.description }}</p>
-            <button
-              @click="useItem"
-              class="button"
-              :class="canUseChosenItem ? '' : 'disabled'"
-            >
-              Use
-            </button>
-          </div>
-        </div>
-        <button class="button" @click="closeItem">{{ '<--' }}</button>
-      </div>
-      <div v-else>
-        <h2>The inventory is empty!</h2>
-      </div>
-    </template>
-  </modal>
+  <div v-if="!chosenItem && Object.keys(itemsToDisplay).length > 0">
+    <InventorySection
+      v-for="section in sections"
+      :key="section.id"
+      :items="section.items"
+      :title="section.title"
+      :id="section.id"
+      @chosen="clickItem"
+    />
+  </div>
+  <ItemDetails
+    v-else-if="typeof chosenId === 'string'"
+    :item="items[chosenId]"
+    @use="useItem"
+    @close="closeItem"
+  />
+  <div v-else>
+    <h2>The inventory is empty!</h2>
+  </div>
 </template>
 
 <script lang="ts">
-import { getAssetUrl, getConfig, ItemData } from '@/config';
+import { getConfig, getItemConfig, ItemData } from '@/config';
 import { useDialogStore } from '@/stores/dialog-store';
 import { useInventory, ItemState } from '@/stores/inventory-store';
 import { useVM } from '@/stores/vm-store';
 import { audioEvent } from '@/utils/audio-loader';
 import { error } from '@/utils/error-handling';
 import { computed, defineComponent } from 'vue';
-import Modal from './utils/modal-window.vue';
+import InventorySection, {
+  InventorySectionProps,
+} from './inventory/inventory-section.vue';
+import ItemDetails from './inventory/item-details.vue';
 
 export default defineComponent({
   setup() {
@@ -65,9 +41,7 @@ export default defineComponent({
     const currentlyChoosing = computed(() => dialogStore.currentDialog.choices);
     return { items, currentlyChoosing };
   },
-  components: {
-    Modal,
-  },
+  emits: ['close'],
   data() {
     return {
       chosenId: false as string | false,
@@ -77,17 +51,6 @@ export default defineComponent({
   methods: {
     close() {
       this.$emit('close');
-    },
-    getItemStyle(item: string): any {
-      return {
-        backgroundImage: `url(${getAssetUrl(this.itemConf[item].icon)})`,
-      };
-    },
-    getItemName(item: string): string {
-      return this.itemConf[item].name;
-    },
-    getItemConf(item: string) {
-      return this.itemConf[item];
     },
     clickItem(item: string) {
       this.chosenId = item;
@@ -109,28 +72,10 @@ export default defineComponent({
         }
       }
     },
-    canUseItem(item: ItemState) {
-      const conf = this.itemConf[item.id];
-      if (
-        conf &&
-        conf.onUse &&
-        !useInventory().isInteractionTagBlocked(conf.tag) &&
-        !this.currentlyChoosing
-      ) {
-        return true;
-      }
-      return false;
-    },
   },
   computed: {
-    itemsToDisplay(): { [key: string]: ItemState } {
-      const items: { [key: string]: ItemState } = {};
-      for (const key in this.items) {
-        if (this.items[key].amount > 0) {
-          items[key] = this.items[key];
-        }
-      }
-      return items;
+    itemsToDisplay(): ItemState[] {
+      return Object.values(this.items);
     },
     chosenItem(): null | ItemState {
       if (this.chosenId) {
@@ -139,7 +84,7 @@ export default defineComponent({
       return null;
     },
     canUseChosenItem() {
-      return this.canUseItem(this.chosenItem!);
+      return useInventory().canUseItem(this.chosenItem!);
     },
     chosenItemConf(): null | ItemData {
       if (this.chosenId) {
@@ -147,10 +92,38 @@ export default defineComponent({
       }
       return null;
     },
-    itemConf(): { [key: string]: ItemData } {
-      return getConfig().items;
+    itemConf(): {
+      [key: string]: ItemData;
+    } {
+      return getConfig().items.items;
+    },
+    sections(): InventorySectionProps[] {
+      // Split the inventory into sections based on the item category, with a default category for items not in any.
+      const categories = getConfig().items.categories;
+      const sections: InventorySectionProps[] = [];
+      const possibleSections = this.itemsToDisplay.reduce((acc, item) => {
+        const category = getItemConfig(item.id).category ?? 'default';
+        const categoryConfig = categories.find((c) => c.id === category);
+        if (!categoryConfig) {
+          error(`Unknown category ${category}`);
+          return acc;
+        }
+        let matchingSection = acc.find((s) => s.id === category);
+        if (!matchingSection) {
+          matchingSection = {
+            id: category,
+            title: categoryConfig.title,
+            items: [],
+          };
+          acc.push(matchingSection);
+        }
+        matchingSection.items.push(item);
+        return acc;
+      }, sections);
+      return possibleSections;
     },
   },
+  components: { InventorySection, ItemDetails },
 });
 </script>
 
@@ -164,35 +137,6 @@ export default defineComponent({
   grid-auto-rows: auto;
   grid-template-columns: repeat(3, 1fr);
   grid-gap: 20px 20px;
-}
-
-.item-display {
-  width: 200px;
-  height: 300px;
-  position: relative;
-  background-repeat: no-repeat;
-  background-size: contain;
-}
-
-.item-title {
-  position: absolute;
-  bottom: 0px;
-  text-align: center;
-  width: 100%;
-  color: var(--inventory-text-color);
-  background: var(--inventory-text-background);
-}
-
-.item-amount {
-  position: absolute;
-  top: 0;
-  right: 0;
-  font-weight: 700;
-  font-size: 1.25rem;
-  color: var(--inventory-amount-color);
-  width: 40px;
-  height: 40px;
-  background-color: var(--inventory-amount-background);
 }
 
 .item-description-container {