+++ packages/narrat/src/menu-buttons/menu-buttons.ts
@@ -2,7 +2,8 @@ import MainMenu from '../components/MainMenu.vue';
 import InventoryUi from '../components/inventory-ui.vue';
 import QuestsUi from '../components/quests-ui.vue';
 import Skills from '../components/SkillsWindow.vue';
-import { MenuButtonState, useMenu } from '@/stores/menu-store';
+
+import { MenuState, MenuTabState, useMenu } from '@/stores/menu-store';
 import { vm } from '@/vm/vm';
 
 export function registerDefaultMenuButtons(app: any) {
@@ -11,48 +12,80 @@ export function registerDefaultMenuButtons(app: any) {
   app.component('InventoryUi', InventoryUi);
   app.component('QuestsUi', QuestsUi);
   app.component('SkillsWindow', Skills);
-  useMenu().buttons = [
-    {
+  useMenu().menus = {
+    system: {
+      label: 'System',
+      cssClass: 'system-menu',
+      id: 'system',
+      tabs: [
+        {
+          id: 'system-tab',
+          cssId: 'system-tab',
+          text: 'System',
+          component: 'MainMenu',
+        },
+      ],
+      activeTab: 0,
+    },
+    menu: {
+      label: 'Menu',
       id: 'menu',
-      cssId: 'menu-button',
-      text: 'Menu',
-      component: 'MainMenu',
+      cssClass: 'menu-menu',
+      tabs: [
+        {
+          id: 'skills',
+          cssId: 'skills-menu-button',
+          text: 'Skills',
+          condition: () => useMenu().showSkills,
+          component: 'SkillsWindow',
+        },
+        {
+          id: 'inventory',
+          cssId: 'inventory-menu-button',
+          text: 'Items',
+          condition: () => useMenu().showInventory,
+          component: 'InventoryUi',
+        },
+        {
+          id: 'quests',
+          cssId: 'quests-menu-button',
+          text: 'Quests',
+          condition: () => useMenu().showQuests,
+          component: 'QuestsUi',
+        },
+      ],
+      activeTab: 0,
     },
-    {
-      id: 'skills',
-      cssId: 'skills-menu-button',
-      text: 'Skills',
-      condition: () => useMenu().showSkills,
-      component: 'SkillsWindow',
-    },
-    {
-      id: 'inventory',
-      cssId: 'inventory-menu-button',
-      text: 'Items',
-      condition: () => useMenu().showInventory,
-      component: 'InventoryUi',
-    },
-    {
-      id: 'quests',
-      cssId: 'quests-menu-button',
-      text: 'Quests',
-      condition: () => useMenu().showQuests,
-      component: 'QuestsUi',
-    },
-  ];
+  };
 }
 
-export function registerMenuButton(config: MenuButtonState, component: any) {
+export function registerMenuButton(
+  id: string,
+  config: MenuState,
+  components: { [key: string]: any },
+) {
+  const app = (window as any).narrat.app;
+  for (const tab in config.tabs) {
+    const componentId = config.tabs[tab].component;
+    app.component(componentId, components[componentId]);
+  }
+  useMenu().addMenu(id, config);
+}
+export function registerMenuTab(
+  id: string,
+  config: MenuTabState,
+  component: any,
+) {
   const app = (window as any).narrat.app;
   app.component(config.component, component);
-  useMenu().addMenuOption(config);
+  useMenu().addMenuOption(id, config);
 }
 
 export function addMenuButtonsFromPlugins() {
   for (const plugin of vm.plugins) {
     if (plugin.customMenuButtons) {
       for (const button of plugin.customMenuButtons) {
-        registerMenuButton(button.config, button.component);
+        registerMenuButton(button.config.id, button.config, button.component);
       }
     }
   }